DCB-306: Complete Material UI switchover
This change removes react-bootstrap and other redundant packages. It
also implements an adaptor for the MUI and Next Link components, to
improve performance, and removes redundant components. There are also
some small accessibility fixes and tidying up in this change: the result
is zero accessibility issues in the AXE Dev Tools.