fix: Add aria-label to breadcrumbs [DCB-878]
This is a small accessibility improvement to bring our Breadcrumbs
implementation in line with MUI and WAI-ARIA best practice. It just adds
an aria-label to the component to identify it clearly as a breadcrumb
trail and a navigation landmark - this makes it easier to locate.

See mui.com/material-ui/react-breadcrumbs/#accessibility