feat: Introduce ErrorLink to Apollo Client implementation [DCB-1154]
This introduces the Error Link for better error handling in Apollo
Client.
This commit is a very basic implementation, originally intended to
simply log errors. It will be extended further alongside the RetryLink
to improve recovery from an error stat
This introduces the Error Link for better error handling in Apollo
Client.
This commit is a very basic implementation, originally intended to
simply log errors. It will be extended further alongside the RetryLink
to improve recovery from an error state.