fix: Restore behaviour of return to patron request button [DCB-1206]
This change makes the 'return to patron request' button once again jump
back to the 'audit log' section of the patron request page.

Oddly, the fix for this was putting the Apollo ErrorLink in its own file
and importing it - something in its initialisation appeared to be
causing the behaviour. The audit log key itself has also been changed to
auditlog to avoid any translation issues.