fix: Display 'not set' participation status in library grids
Previously, library grids were displaying nothing when a library's
participation status was 'Not set'. This gives the user the impression
that the library in question doesn't have that option (i.e. has no
associated agency), when in fact it does. Given that libraries start by
default with participation set to null, it is important to ensure users
are aware of the 'not set' status.