fix: Restrict access to the data change log [DCB-1302]
This change introduces the concept of admin-only routes. These routes
are routes that can only be made available to users with certain roles,
and are defined along with the role groupings in middleware.ts.

The first of these routes is the /serviceInfo/dataChangeLog route and
all child routes (i.e. individual data change log pages). Note also that
the UI navigation to the dataChangeLog page will only be visible for
authorised users.