chore: Hide actions column on mappings page for non-admins
Also includes some refactoring of data grid and associated functions to
enable this and to make it easier to maintain. Further refactoring is
left for DCB-152chore: Hide actions column on mappings page for
non-admins

Also includes some refactoring of data grid and associated functions to
enable this and to make it easier to maintain. Further refactoring is
not in the scope of this ticket and is left for DCB-1523.