fix: Upgrade MUI-X, apply manual casting for useGridApiContext [DCB-1816]
There is an issue (16320) in mui-x that is causing us to see type errors
when using useGridApiContext() and useGridApiRef(). This is happening
because of behaviour introduced to handle React 19 changes inadvertently
affecting us also. By upgrading to latest mui-x, we fix the issue for
useGridApiRef(), and the manual casting can be used to handle the issue
for useGridApiContext() until the mui-x team release a fix.