fix: Fix deleted mappings showing up on library mapping pages [DCB-1819]
The queries for the library mappings data grids had incorrectly copied
the preset queries for the main data grids, resulting in brackets being
where they shouldn't be, which caused the expression to evaluate
incorrectly.
I have removed those brackets and standardised the grids to use AND NOT
deleted:true. When we no longer have to support the possibility of
deleted being null, we can simplify this further.
The 'main' mapping grids should stay the same, as their preset queries
are working correctly and need to be formatted differently, as they only
have two conditions.