chore: Re-work Apollo Client handling [DCB-1440]
This commit re-works the apollo client creation and handling, by
introducing the apollo-client.js file which will now handle the creation
of the Apollo Client for the application, replacing the
ApolloProviderWrapper. This means we can then use the default
ApolloProvider in app.tsx and pass the created client to it. The aim of
this is to simplify the implementation, reduce the re-renders that were
caused by sometimes having multiple clients, and improve Apollo
stability.

It also increases the token refresh time to 15 minutes, which will need
to be reflected across all Keycloak environments.