chore: Add custom stepper components [DCB-1555]
Adds the DCBStepIcon component to deliver a customised stepper UX for
the expedited checkout workflow, the StatusStepConnector component for a
status-aware step connector, and separates out the three steps into
their own component files.

Also establishes a status constant for statuses which should trigger a
request success, and helper functions for applying step stylings.

Adds a menu to cluster.tsx for placing a staff request and checking out
an item.