chore: UX improvements to library, consortium, location forms [DCB-1503]
Removes the EditableAttribute component and moves the library,
consortium and location forms over to using react-hook-form instead.

Also does a bit of refactoring to try and get all the handleTabChange
functions in one place, and for the editAndDeleteActions implementations
of handleSaveConfirmation, handleEdit, and handleCancellation to be used
in all 3 forms.