feat: Support for a future live diagnostics page [DCB-1915]
This commit contains support for a future page that will show key
diagnostic information to consortium administrators, helping them when
contacting support. It will also assist developers and implementation
partners in identifying issues.

While this page is not yet ready, this commit introduces a basic, not
publicly accessible page as the backbone of future support for this
feature.