feat: Display 'Items not shown' for comparison [DCB-1947]
Adds a display for items that are not currently being shown to patrons
in the Items view. This allows admins to review these items and check if
they are intentionally not being shown, or if there are configuration
issues preventing them from being displayed to patrons.
Also included in DCB Admin for Libraries.