feat: Add library bibs section [DCB-2052]
To aid in debugging clustering issues. This is a 'first pass' of a bib
records page for the individual libraries, showing all of the bib
records for a given library.
The association is made between the source system ID on the bib record
and the associated Host LMS on the library. For libraries with more than
one Host LMS, an OR query is used with the two IDs.
Sorting is disabled on this grid to avoid the expensive operation of
sorting millions of bib records. There is also the option of disabling
filters entirely if this page causes issues.